semi-final fraud prover setup
