+++ omgx_utilities/fraud-prover/package.json
@@ -39,6 +39,7 @@
     "bcfg": "^0.1.6",
     "dotenv": "^8.2.0",
     "hardhat": "^2.2.1",
+    "hardhat-deploy": "^0.7.5",
     "merkle-patricia-tree": "^4.2.0",
     "merkletreejs": "^0.2.18",
     "rlp": "^2.2.6",
